convert start and version commands
